[Ref/#208] 레거시 모델 제거, 모델명&함수명 수정, 파일 분리
