+++ ILSANG/Sources/Model/Rank.swift
@@ -1,0 +1,13 @@
+//
+//  Rank.swift
+//  ILSANG
+//
+//  Created by Lee Jinhee on 12/11/24.
+//
+
+struct Rank: Decodable {
+    let xpType: String
+    let xpPoint: Int
+    let customerId: String
+    let nickname: String
+}
