+++ ILSANG/Sources/Network/UserNetwork.swift
@@ -29,7 +29,7 @@ final class UserNetwork {
         }
     }
     
-    func getUserRank(xpstat: String) async -> Result<UserRank, Error> {
+    func getUserRank(xpstat: String) async -> Result<Response<[Rank]>, Error> {
         let parameters: Parameters = ["xpType":xpstat, "size": 20]
         
         return await Network.requestData(url: rankUrl, method: .get, parameters: parameters, withToken: true)
